ccnx: Handle requests in separate processes.
Because several git functions terminate the current process when they
encounter an error, we need to handle requests for refs and objects into
a particular repository in a separate child process. When the child
exists, the parent process can simply re-create it when a new request
for the same repository is received.

Furthermore, once a process initializes access to a particular
repository, it is difficult to switch to another git repository in the
same process, because the initialization functions setup many global
variables. To deal with this situation, we create and maintain one
process for each git repository.

The parent(dispatcher) process registers ccnx filter for a prefix common
for all git repositories (typically their parent directory). When a new
Interest request arrives, the parent process extracts the name of the
desired git repository from the request and creates a new child process.

The child process then registers a new (longer) filter for this
particular git repository only. As long as the child process is alive
and maintains its link to ccnd and the filter registration, further
Interest requests for that particular git repository are delivered to
the child process rather than to the parent. Its filter is longer than
the common filter of the parent process and thus wins the longest prefix
match.

The parent process keeps receiving Interest packets only for git
repositories that have no child process yet. Such repositories will only
be matched by the filter for the common prefix, no longer filter exists
yet.

For example, if all git repositories handled by the daemon are in
"/git", then the parent process registers ccnx filter for the prefix
"/git". When an Interest packet with prefix /git/ccnx arrives, it will
be initially delivered to the parent process. The parent process creates
a child which registers a filter for prefix "/git/ccnx" immediately.
Further Interest packets for "/git/ccnx" will be delivered to the child
process because its filter is longer and matches more prefix components.