ccnx: Generate list of all refs in a repository.
Generate a Data packet containing the list of all refs (branches and
tags) available in a git repository in response to Interest packets with
prefix /git/<repo_name>/refs. The function uses its own binary format:

  struct ref_entry {
      unsigned int flags;
      unsigned int ref_len;
      unsigned int symref_len;
      unsigned char sha1[20];
      char ref_name[ref_len + 1];
      char symref_target[symref_len + 1];
  };

All integer fields are in network byte order. The flags field contains
encodes ref related flags such as REF_ISSYMREF and REF_ISPACKED. The
SHA1 is stored in binary form. The field symref_len contains number of
characters in symref_target. This is currently only used for HEAD, in
all other cases it will be set to 0. Both strings are terminated with a
zero for convenience, but neither ref_len nor symref_len contain the
zero terminator characters.

Tag refs are expanded, we add another ref for them with ^{} suffix which
points to the first non-tag object. Such refs are used by clients to
synchronize remote tags.

Additionally we also append HEAD as a refs to the list so that clients
do not have to ask for it separately.