ccnx: Upload compressed git objects to ccn.
Individual objects are stored compressed with zlib in the git
repository. We get their body uncompressed from git and before
submitting individual git objects into ccn content store, we recreate
the object in exactly the same way it is stored in git. That is, we add
the object's header at the beginning of the body and compress the whole
buffer with zlib.