ccnx: Split Data packets into multiple fragments.
If the size of a Data packet exceeds a certain pre-configured threshold,
split it into multiple fragments and submit them into the ccnd content
store individually. Each such fragment has its fragment number counted
from 0 in the last component of its name. Each fragment also has the
number of the last fragment in the series in SignedInfo/FinalBlockID.

To be ready to receive Interests for individual fragments (with longer
prefix name), we need to update all expressions that compare the value
of comps_left in child_handler.

We always submit all fragments for a particular object, even if we
receive an Interest for just one particular fragment.

The maximum number of a Data packet can be configured with the new -s
command line option. Data packets larger than the value will be
automatically fragmented.