ccnx: Support for Data packets with fragments.
If a git object is too big and does not fit into a single Data packet,
the daemon generating Data packets from the content of a git repository
splits it into multiple fragments and sends those fragment to the ccn
network individually.

The client inspects the last component of Data's name and the value of
of FinalBlockID. If they indicate that what we've received is on a
fragment of the whole object, it fetches all remaining fragments, one
after another, to obtain the whole git object.